Overhauled gridReducer to recalculate availableValues
Made GridState and CellState completely immutable.
Simplified the logic significantly.

Modified the algorithm so that when a cell is modified, the affected
cells get their availableValues completely refreshed instead of just
adding or removing values. This fixes the issue where if the
availableValues get corrupted if you start entering wrong values into
cells,