Disable mouse scrolling on the input cell in focus
Mouse scrolling on the selected input cell causes it to
increment/decrement the cell value. Changing the values with scrolling
is most probably unintended.

The scrolling also has a side-effect of totally messing up the
availableValues of other cells. This is because the related cells
get elimination applied to their availableValues, which can't be
restored after